requirements.txt
click
dateparser
faker
pyyaml
tqdm